refactor: consolidate ValidationResult into @taquito/core
re #1994
