+++ packages/taquito-core/src/enums.ts
@@ -1,0 +1,6 @@
+export enum ValidationResult {
+  NO_PREFIX_MATCHED,
+  INVALID_CHECKSUM,
+  INVALID_LENGTH,
+  VALID,
+}
