+++ packages/taquito-michelson-encoder/src/tokens/chain-id.ts
@@ -5,8 +5,9 @@ import {
   TokenValidationError,
   SemanticEncoding,
 } from './token';
-import { validateChain, ValidationResult } from '@taquito/utils';
+import { validateChain } from '@taquito/utils';
 import { BaseTokenSchema } from '../schema/types';
+import { ValidationResult } from '@taquito/core';
 
 export class ChainIDValidationError extends TokenValidationError {
   name = 'ChainIDValidationError';
