+++ packages/taquito-rpc/src/taquito-rpc.ts
@@ -64,8 +64,8 @@ import {
   OriginationProofParams,
 } from './types';
 import { castToBigNumber } from './utils/utils';
-import { validateAddress, validateContractAddress, ValidationResult } from '@taquito/utils';
-import { InvalidAddressError } from '@taquito/core';
+import { validateAddress, validateContractAddress } from '@taquito/utils';
+import { InvalidAddressError, ValidationResult } from '@taquito/core';
 
 export { castToBigNumber } from './utils/utils';
