+++ packages/taquito/src/tz/rpc-tz-provider.ts
@@ -5,13 +5,8 @@ import { Operation } from '../operations/operations';
 import { RPCActivateOperation } from '../operations/types';
 import { TzProvider } from './interface';
 import { OpKind } from '@taquito/rpc';
-import {
-  validateAddress,
-  ValidationResult,
-  validateKeyHash,
-  InvalidKeyHashError,
-} from '@taquito/utils';
-import { InvalidAddressError } from '@taquito/core';
+import { validateAddress, validateKeyHash, InvalidKeyHashError } from '@taquito/utils';
+import { InvalidAddressError, ValidationResult } from '@taquito/core';
 
 export class RpcTzProvider extends OperationEmitter implements TzProvider {
   constructor(context: Context) {
